refactor: extract out label-modal to reuse in editor
