+++ src/Components/Aside/Aside.jsx
@@ -1,53 +1,23 @@
 import { useState } from 'react';
 import { NavLink } from 'react-router-dom';
-import { Backdrop } from '../Backdrop/Backdrop';
+import { LabelModal } from '../../Components';
 import { useLabel } from '../../Context';
 import styles from './Aside.module.css';
 
 const activeClass = ({ isActive }) => (isActive ? styles.active : '');
 
 const Aside = () => {
-	const [newLabel, setNewLabel] = useState('');
 	const [labelModal, setLabelModal] = useState(false);
 
 	const {
 		labelState: { labels },
-		labelDispatch,
 	} = useLabel();
-
-	const newLabelHandler = () => {
-		if (labels.some((label) => label === newLabel)) {
-			console.log('label already present');
-			setLabelModal(false);
-			setNewLabel('');
-		} else {
-			labelDispatch({ type: 'ADD_NEW_NOTE', payload: newLabel });
-			setLabelModal(false);
-			setNewLabel('');
-		}
-	};
 
 	return (
 		<aside className={styles.aside}>
 			{labelModal ? (
 				<>
-					<Backdrop setLabelModal={setLabelModal} />
-					<div className={styles.labelModal}>
-						<div className="input-container">
-							<label htmlFor="label">Add new Label</label>
-							<input
-								type="text"
-								id="label"
-								name="label"
-								placeholder="Travel"
-								value={newLabel}
-								onChange={(e) => setNewLabel(e.target.value)}
-							/>
-						</div>
-						<button className="btn btn-info btn-round" onClick={newLabelHandler}>
-							Add
-						</button>
-					</div>
+					<LabelModal setLabelModal={setLabelModal} />
 				</>
 			) : null}
 			<div className={styles.asideBtnList}>
